+++ python_plus/python_plus/__init__.py
@@ -10,7 +10,7 @@ from . import scripts
 __title__ = 'python_plus'
 __author__ = 'Antonio Maria Vigliotti'
 __copyright__ = 'Copyright 2018-2024 SHS-AV srl'
-__version__ = '2.0.13'
+__version__ = '2.0.14'
 
 PYCODESET = 'utf-8'
 if PY3:
@@ -288,6 +288,7 @@ class __(object, with_metaclass(Base__, text_type)):
             enquote=enquote,
             strip=strip,
         )
+