+++ z0lib/z0lib/scripts/main.py
@@ -22,7 +22,7 @@ import gzip
 import shutil
 
 
-__version__ = "2.0.9"
+__version__ = "2.0.10"
 
 
 def get_pypi_info(pkgname):
@@ -226,4 +226,3 @@ def main(cli_args=None):
         # copy_pkg_data(setup_args, verbose)
         copy_pkg_data(pypi_metadata, verbose)
     return 0
-