+++ zerobug/zerobug/docs/conf.py
@@ -25,9 +25,9 @@ copyright = "2019-23, SHS-AV s.r.l."
 author = "Antonio Maria Vigliotti"
 
 # The short X.Y version
-version = "2.0.16"
+version = "2.0.17"
 # The full version, including alpha/beta/rc tags
-release = "2.0.16"
+release = "2.0.17"
 
 
 # -- General configuration ---------------------------------------------------
@@ -170,6 +170,7 @@ autodoc_default_options = {
     "undoc-members": True,
     "exclude-members": "Macro, SanityTest",
 }
+